Make git fetching static prop on buildtime
