+++ lib/gitlab_contribution_converter.js
@@ -1,15 +1,11 @@
 const moment = require('moment')
 const fs = require('fs')
-const axios = require('axios')
+// const axios = require('axios')
 const dotenv = require('dotenv')
 dotenv.config()
+
 const gitLabApiClient = require('./gitlabApiClient.js')
 
-gitLabApiClient.init(process.argv[2]) // init client with [1,2] num from args
-
-const since_date = moment().add('-400', 'days')
-
-let all_events_cache = []
 try {
   all_events_cache = JSON.parse(
     fs.readFileSync(
@@ -18,7 +14,7 @@ try {
 
 }
 
-const main = async () => {
+export default async () => {
   if (all_events_cache.length < 1 || moment(all_events_cache[0].created_at) < moment().add('-30', 'days')) { // cache empty or too old
     console.log('Existing cache empty or most recent event to long ago, loading big event data now...')
     all_events_cache = await gitLabApiClient.getAll(
@@ -52,13 +48,13 @@ const main = async () => {
       },
     }
   })
+  const result = {
+    date: moment().toISOString(),
+    data: filtered,
+  }
   fs.writeFileSync(
     `lib/gitlab_contribution_history_cache_${gitLabApiClient.config.GITLAB_INSTANCE}.json`,
-    JSON.stringify({
-      date: moment().toISOString(),
-      data: filtered,
-    }))
+    JSON.stringify(result))
   console.log('done.')
+  return filtered
 }
-main()
-
